Update dependencies.
Update for compose, gradle + agp and kotlin compiler for compose. This introduced a viewbinding bug, where included <merge> layouts throw a NPE in runtime. (https://issuetracker.google.com/issues/175741240). Fixed this by using LinearLayout as the root, but now the app name is included and the the content is centered.